# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SkyRadar REST API
  version: 1.0.0
  description: |
    SkyRadar 最终 RESTful API 契约源。所有业务接口使用 `/api/v1/*` 路径。

    成功响应统一使用 `data`、`meta`、`links`；错误响应统一使用
    `error`、`message`、`detail`、`request_id`。
servers:
  - url: /
...
paths:
...
  /api/v1/leakages:
    get:
      tags:
        - Leakages
      summary: 获取泄露结果列表
      operationId: listLeakages
      parameters:
        - name: security
          in: query
          required: false
          description: 安全状态过滤，0 表示待确认风险，1 表示安全或忽略。
          schema:
            type: integer
            enum:
              - 0
              - 1
          example: 0
        - name: ignored
          in: query
          required: false
          description: 是否忽略。
          schema:
            type: boolean
          example: false
        - name: reviewed
          in: query
          required: false
          description: 是否已有处理备注。
          schema:
            type: boolean
          example: true
        - name: tag
          in: query
          required: false
          description: 规则标签过滤。
          schema:
            type: string
          example: github-token
        - name: language
          in: query
          required: false
          description: 语言过滤。
          schema:
            type: string
          example: Python
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 泄露结果分页列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeakageListSuccess"
              example:
                data:
                  - _id: leak-001
                    link: https://github.com/example/repo/blob/main/config.py
                    project: example/repo
                    project_url: https://github.com/example/repo
                    language: Python
                    username: example
                    avatar_url: https://avatars.githubusercontent.com/u/1?v=4
                    filepath: config.py
                    filename: config.py
                    security: 0
                    ignore: 0
                    tag: github-token
                    desc: ""
                    datetime: "2026-06-15T10:00:00Z"
                    discovered_at: "2026-06-15T10:05:00Z"
                    discovered_timestamp: 1781517900
                meta:
                  page: 1
                  page_size: 20
                  total: 1
                links:
                  self: /api/v1/leakages?page=1&page_size=20
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/leakages/{leakage_id}:
    get:
      tags:
        - Leakages
      summary: 获取泄露详情元信息
      operationId: getLeakage
      parameters:
        - $ref: "#/components/parameters/LeakageId"
      responses:
        "200":
          description: 泄露详情元信息，不包含代码正文。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeakageSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags:
        - Leakages
      summary: 更新泄露处理结果
      operationId: updateLeakage
      parameters:
        - $ref: "#/components/parameters/LeakageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeakageUpdateRequest"
            example:
              security: 0
              ignored: false
              desc: 确认为有效泄露
              project: example/repo
      responses:
        "200":
          description: 泄露处理结果已更新。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeakageSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/leakages/{leakage_id}/code:
    get:
      tags:
        - Leakages
      summary: 获取泄露代码和受影响资产
      operationId: getLeakageCode
      parameters:
        - $ref: "#/components/parameters/LeakageId"
      responses:
        "200":
          description: 泄露代码和受影响资产。代码内容为 base64 编码字符串。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeakageCodeSuccess"
              example:
                data:
                  code: U0tZX1NBTVBMRV9LRVk=
                  affect:
                    - type: file
                      value: config.py
                meta: {}
                links:
                  self: /api/v1/leakages/leak-001/code
                  leakage: /api/v1/leakages/leak-001
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/leakages/{leakage_id}/ai-analysis:
    post:
      tags:
        - Leakages
      summary: 重新投递单条泄露 AI 分析
      description: 仅提交异步分析任务；AI 分析失败不会影响泄露结果保存或通知链路。
      operationId: createLeakageAIAnalysis
      parameters:
        - $ref: "#/components/parameters/LeakageId"
      responses:
        "202":
          description: 单条 AI 分析任务已提交。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeakageAIAnalysisSuccess"
              example:
                data:
                  id: leak-001
                  ai_analysis:
                    status: pending
                meta: {}
                links: {}
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    LeakageId:
      name: leakage_id
      in: path
      required: true
      schema:
        type: string
      example: leak-001
...
  responses:
    BadRequest:
      description: 请求格式或参数无法处理。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: 资源不存在。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: 请求语义校验失败。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ApiSuccess:
      type: object
      required:
        - data
      properties:
        data:
          nullable: true
        meta:
          type: object
          additionalProperties: true
        links:
          type: object
          additionalProperties:
            type: string
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          pattern: "^[a-z0-9_]+$"
          example: invalid_request
        message:
          type: string
          example: 请求参数无效
        detail:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: string
          example:
            field: page_size
            reason: must be between 1 and 100
        request_id:
          type: string
          nullable: true
          example: req_01HZYXAMPLE
    PageMeta:
      type: object
      required:
        - page
        - page_size
        - total
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    LeakageListSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Leakage"
            meta:
              $ref: "#/components/schemas/PageMeta"
    LeakageSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Leakage"
    Leakage:
      type: object
      required:
        - _id
        - link
        - project
        - security
        - ignore
        - tag
        - datetime
        - discovered_at
        - discovered_timestamp
      properties:
        _id:
          type: string
        link:
          type: string
          format: uri
        project:
          type: string
        project_url:
          type: string
          format: uri
        language:
          type: string
        username:
          type: string
        avatar_url:
          type: string
          format: uri
        filepath:
          type: string
        filename:
          type: string
        security:
          type: integer
          enum:
            - 0
            - 1
        ignore:
          type: integer
          enum:
            - 0
            - 1
        tag:
          type: string
        desc:
          type: string
        datetime:
          type: string
          description: GitHub 侧文件或仓库更新时间。
        discovered_at:
          type: string
          description: SkyRadar 首次写入该泄露结果的发现时间。
        discovered_timestamp:
          type: number
          description: SkyRadar 发现时间的 Unix timestamp。
        ai_analysis:
          $ref: "#/components/schemas/AIAnalysis"
    LeakageUpdateRequest:
      type: object
      required:
        - security
      properties:
        project:
          type: string
        security:
          type: integer
          enum:
            - 0
            - 1
        ignored:
          type: boolean
        desc:
          type: string
    LeakageCodeSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/LeakageCode"
    LeakageCode:
      type: object
      required:
        - code
        - affect
      properties:
        code:
          type: string
          description: 代码内容。
        affect:
          type: array
          items:
            $ref: "#/components/schemas/AffectedAsset"
    LeakageAIAnalysisSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              type: object
              required:
                - id
                - ai_analysis
              properties:
                id:
                  type: string
                ai_analysis:
                  $ref: "#/components/schemas/AIAnalysis"
    AIAnalysis:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - pending
            - running
            - success
            - failed
            - skipped
        risk_level:
          type: string
          enum:
            - low
            - medium
            - high
            - unknown
        summary:
          type: string
        evidence:
          type: array
          items:
            type: string
        recommendation:
          type: string
        false_positive_reason:
          type: string
        is_useful:
          type: boolean
          description: AI 是否判断该项目或命中结果值得推送 webhook。
        usefulness_reason:
          type: string
        matched_interests:
          type: array
          items:
            type: string
        model:
          type: string
        analyzed_at:
          type: string
          format: date-time
        error:
          type: string
    AffectedAsset:
      type: object
      additionalProperties: true
      properties:
        type:
          type: string
        value:
          type: string
...